src/main/java/chess/domain/Board.java
package chess.domain;

import chess.domain.piece.Empty;
import chess.domain.piece.Piece;
import java.util.HashMap;
import java.util.Map;

public class Board {
    
    private final Map<Position, Piece> board;
    
    private Board(final Map<Position, Piece> board) {
        this.board = board;
    }
    
    
    public static Board create() {
        Map<Position, Piece> board = new HashMap<>();
        for (File file : File.values()) {
            for (Rank rank : Rank.values()) {
                Position position = Position.from(file.getLabel() + rank.getLabel());
                board.put(position, Empty.create());
            }
        }
        return new Board(board);
    }
}
